+++ sentry-kotlin-multiplatform/src/appleMain/kotlin/io/sentry/kotlin/multiplatform/TypeAliases.apple.kt
@@ -4,13 +4,13 @@ import cocoapods.Sentry.SentryAttachment
 import cocoapods.Sentry.SentryBreadcrumb
 import cocoapods.Sentry.SentryEvent
 import cocoapods.Sentry.SentryException
+import cocoapods.Sentry.SentryFeedback
 import cocoapods.Sentry.SentryId
 import cocoapods.Sentry.SentryLevel
 import cocoapods.Sentry.SentryMessage
 import cocoapods.Sentry.SentryOptions
 import cocoapods.Sentry.SentryScope
 import cocoapods.Sentry.SentryUser
-import cocoapods.Sentry.SentryUserFeedback
 
 internal typealias CocoaUser = SentryUser
 internal typealias CocoaBreadcrumb = SentryBreadcrumb
@@ -19,7 +19,7 @@ internal typealias CocoaScope = SentryScope
 internal typealias CocoaSentryId = SentryId
 internal typealias CocoaSentryLevel = SentryLevel
 internal typealias CocoaAttachment = SentryAttachment
-internal typealias CocoaUserFeedback = SentryUserFeedback
+internal typealias CocoaUserFeedback = SentryFeedback
 internal typealias CocoaSentryEvent = SentryEvent
 internal typealias CocoaMessage = SentryMessage
 internal typealias CocoaSentryException = SentryException