+++ sentry-kotlin-multiplatform/src/appleTest/kotlin/io/sentry/kotlin/multiplatform/BreadcrumbTestConverter.kt
@@ -7,13 +7,13 @@ import io.sentry.kotlin.multiplatform.protocol.Breadcrumb
 actual data class BreadcrumbTestConverter actual constructor(
     val breadcrumb: Breadcrumb,
 ) {
-    actual fun getType(): String? = breadcrumb.toCocoaBreadcrumb().type
+    actual fun getType(): String? = breadcrumb.toCocoaBreadcrumb().type()
 
-    actual fun getCategory(): String? = breadcrumb.toCocoaBreadcrumb().category
+    actual fun getCategory(): String? = breadcrumb.toCocoaBreadcrumb().category()
 
-    actual fun getMessage(): String? = breadcrumb.toCocoaBreadcrumb().message
+    actual fun getMessage(): String? = breadcrumb.toCocoaBreadcrumb().message()
 
-    actual fun getData(): MutableMap<String, Any> = breadcrumb.toCocoaBreadcrumb().data as MutableMap<String, Any>
+    actual fun getData(): MutableMap<String, Any> = (breadcrumb.toCocoaBreadcrumb().data() as Map<String, Any>).toMutableMap()
 
-    actual fun getLevel(): SentryLevel? = breadcrumb.toCocoaBreadcrumb().level.toKmpSentryLevel()
+    actual fun getLevel(): SentryLevel? = breadcrumb.toCocoaBreadcrumb().level().toKmpSentryLevel()
 }